perf: 单独拆分 naive 全局 API 挂载方法
