+++ src/main.ts
@@ -1,6 +1,6 @@
 import './styles/tailwind.css';
 import { createApp } from 'vue';
-import { setupNaive, setupDirectives } from '@/plugins';
+import { setupNaiveDiscreteApi, setupNaive, setupDirectives } from '@/plugins';
 import App from './App.vue';
 import router, { setupRouter } from './router';
 import { setupStore } from '@/store';
@@ -8,8 +8,14 @@ import { setupStore } from '@/store';
 async function bootstrap() {
   const app = createApp(App);
 
+  // 挂载状态管理
+  setupStore(app);
+
   // 注册全局常用的 naive-ui 组件
   setupNaive(app);
+
+  // 挂载 naive-ui 脱离上下文的 Api
+  setupNaiveDiscreteApi();
 
   // 注册全局自定义组件
   //setupCustomComponents();
@@ -20,13 +26,11 @@ async function bootstrap() {
   // 注册全局方法，如：app.config.globalProperties.$message = message
   //setupGlobalMethods(app);
 
-  // 挂载状态管理
-  setupStore(app);
-
   // 挂载路由
-  await setupRouter(app);
+  setupRouter(app);
 
-  // 路由准备就绪后挂载APP实例
+  // 路由准备就绪后挂载 APP 实例
+  // https://router.vuejs.org/api/interfaces/router.html#isready
   await router.isReady();
 
   app.mount('#app', true);
